Add options to SignInHelper
In order to assisted testing in other applications, under scenarios
where the oAuth user will take the place of a relational, persisted,
user, an options hash has been added to the `mock_profile` method.

It also extends the method's flexibility by not `reverse_merging` into
the passed in hash, thus preserving that data.